Re-ogranise tools, and fix known_outcome_agent benchmarking
